% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{calculate_genesetlist_complexity}
\alias{calculate_genesetlist_complexity}
\title{calculate_genesetlist_complexity}
\usage{
calculate_genesetlist_complexity(dataset, species2)
}
\arguments{
\item{dataset}{An XGSA data set created by 'new_XGSA_dataset'}

\item{species2}{Species name in the form 'hsapiens', naming the species to compare the dataset against for gene set complexity.}
}
\value{
This function returns a list of complexity scores, one score for each gene set
}
\description{
This helper function calculates the complexity score of the gene sets in an XGSA data set with respect to a second species. The complexity score is defined in the publication, as the fraction of gene set that has complex homology to another species.
}
\details{
This function calculates the complexity score of the gene sets in an XGSA data set with respect to a second species. The complexity score is defined in the publication, as the fraction of gene set that has complex homology to another species.
}
\examples{
test_mouse_genes <- c("ENSMUSG00000033837", "ENSMUSG00000031965", "ENSMUSG00000053110", "ENSMUSG00000030557", "ENSMUSG00000051159", "ENSMUSG00000040289", "ENSMUSG00000035458", "ENSMUSG00000028780", "ENSMUSG00000062327", "ENSMUSG00000037868")
test_mouse_dataset <- new_XGSA_dataset(species = 'mmusculus', data = list(test_genes = test_mouse_genes), name = 'Test Mouse Genes')
mouse_zebrafish_complexity <- calculate_genesetlist_complexity(test_mouse_dataset,'drerio')
print(mouse_zebrafish_complexity)
}

