% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{check_XGSA_dataset}
\alias{check_XGSA_dataset}
\title{check_XGSA_dataset}
\usage{
check_XGSA_dataset(dataset)
}
\arguments{
\item{dataset}{An XGSA data set}
}
\value{
This function returns a boolean.
}
\description{
This function checks that an XGSA data set is valid
}
\details{
This function checks that an XGSA data set is valid
}
\examples{
test_mouse_genes <- c("ENSMUSG00000033837", "ENSMUSG00000031965", "ENSMUSG00000053110", "ENSMUSG00000030557", "ENSMUSG00000051159", "ENSMUSG00000040289", "ENSMUSG00000035458", "ENSMUSG00000028780", "ENSMUSG00000062327", "ENSMUSG00000037868")
test_mouse_dataset <- new_XGSA_dataset(species = 'mmusculus', data = list(test_genes = test_mouse_genes), name = 'Test Mouse Genes')
check_XGSA_dataset(test_mouse_dataset)
}

