% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{find_supported_datasets}
\alias{find_supported_datasets}
\title{find_supported_datasets}
\usage{
find_supported_datasets(default = TRUE)
}
\value{
This function returns a vector or organism names in the form 'hsapiens'
}
\description{
This function determines which species are currently supported by Ensembl
}
\details{
This function determines which species are currently supported by Ensembl
}
\examples{
supported_species <- find_supported_datasets()
head(supported_species)
}

