% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{generate_best_reciprocal_hits_matrix}
\alias{generate_best_reciprocal_hits_matrix}
\title{generate_best_reciprocal_hits_matrix}
\usage{
generate_best_reciprocal_hits_matrix(hms1, hms2, homology_table)
}
\arguments{
\item{hms1}{Homology matrix with homology scores with species 1 as rows}

\item{hms2}{Homology matrix with homology scores with species 2 as rows}

\item{homology_table}{Table of homology mapping between two species (species 1 in column 1 and species 2 in column 2)}
}
\value{
This function returns a sparse matrix
}
\description{
This function generates a homology matrix based on the best reciprocal hits method.
Note: this can still give not perfect 1-1 mapping if homologs have identical sequence conservation in each direction
}
\details{
This function generates a homology matrix based on the best reciprocal hits method.
Note: this can still give not perfect 1-1 mapping if homologs have identical sequence conservation in each direction
}
\examples{
Not used trivially

}

