% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{get_complex_genes}
\alias{get_complex_genes}
\title{get_complex_genes}
\usage{
get_complex_genes(species1, species2)
}
\arguments{
\item{species1}{Species 1 name in the form 'hsapiens'}

\item{species2}{Species 2 name in the form 'hsapiens'}
}
\value{
This function returns a list with two elements (one for each species), each containing a vector of gene IDs in that species that have more than 1 homolog in the other species
}
\description{
This helper function retrieves complex genes between two species.
}
\details{
This helper function retrieves complex genes between two species.
}
\examples{
human_zebrafish_complex_genes <- get_complex_genes('hsapiens','drerio')
}

