% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{get_ENSEMBL_symbol_map}
\alias{get_ENSEMBL_symbol_map}
\title{get_ENSEMBL_symbol_map}
\usage{
get_ENSEMBL_symbol_map(species)
}
\arguments{
\item{species}{Species name in the form 'hsapiens'}
}
\value{
This function returns a data frame with Ensembl gene IDs in the first column and common gene symbols in the second column.
}
\description{
This helper function retrieves the mapping  between ENSEMBL gene IDs and common gene symbols
}
\details{
This helper function retrieves the mapping  between ENSEMBL gene IDs and common gene symbols
}
\examples{
human_symbol_map <- get_ENSEMBL_symbol_map('hsapiens')
head(human_symbol_map)
}

