% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{get_ENSEMBL_trasncript_table}
\alias{get_ENSEMBL_trasncript_table}
\title{get_ENSEMBL_trasncript_table}
\usage{
get_ENSEMBL_trasncript_table(species)
}
\arguments{
\item{species}{Species name in the form 'hsapiens'}
}
\value{
This helper function returns a data frame with Ensembl gene IDs in the first column and Ensembl transcript IDs in the second column. This can be useful when gene expression data has been mapped to transcripts for example.
}
\description{
This helper function returns a data frame with Ensembl gene IDs in the first column and Ensembl transcript IDs in the second column. This can be useful when gene expression data has been mapped to transcripts for example.
}
\details{
This helper function returns a data frame with Ensembl gene IDs in the first column and Ensembl transcript IDs in the second column. This can be useful when gene expression data has been mapped to transcripts for example.
}
\examples{
transcript_table <- get_ENSEMBL_trasncript_table('hsapiens')
head(transcript_table)

}

