% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{get_GO}
\alias{get_GO}
\title{get_GO}
\usage{
get_GO(species, evidence.codes = c("EXP", "IDA", "IPI", "IMP", "IGI", "IEP",
  "TAS", "IC"), ontologies = c("biological_process", "molecular_function",
  "cellular_component"))
}
\arguments{
\item{species}{Species name in the form 'hsapiens'}

\item{evidence.codes}{A character vector of evidence codes, defaults to 'direct' evidence = c("EXP","IDA","IPI","IMP","IGI","IEP", "TAS", "IC"). A description of GO evidence codes can be found here: http://geneontology.org/page/guide-go-evidence-codes}

\item{ontologies}{A character vector of the desired ontologies to query, defaults to all three of them - i.e. c("biological_process", "molecular_function", "cellular_component")}
}
\value{
This function returns a named list of GO terms with annotated Ensembl gene IDs within each element.
}
\description{
This function retrieves GO mappings with only the supplied evidence codes and from only the supplied ontologies and converts the result into a list. It retrieves the full and redundant GO mappings.
}
\details{
This function retrieves GO mappings with only the supplied evidence codes and from only the supplied ontologies and converts the result into a list. It retrieves the full and redundant GO mappings.
}
\examples{
Human_full_GO <- get_GO('hsapiens')
}

