% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{get_GO_mappings}
\alias{get_GO_mappings}
\title{get_GO_mappings}
\usage{
get_GO_mappings(species)
}
\arguments{
\item{species}{Species name in the form 'hsapiens'}
}
\value{
This function returns a data frame containing Ensembl gene IDs, annotated go terms, GO evidence codes and GO ontology type (BP, MF, CC), in columns 1:4 respectively.
}
\description{
This function retrieves the latest Gene Ontology annotations from Ensembl
}
\details{
This function retrieves the latest Gene Ontology annotations from Ensembl
}
\examples{
human_GO_mappings <- get_GO_mappings('hsapiens')
head(human_GO_mappings)
}

