% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{get_GO_names}
\alias{get_GO_names}
\title{get_GO_names}
\usage{
get_GO_names(species)
}
\arguments{
\item{species}{Species name in the form 'hsapiens'}
}
\value{
This function returns a data frame containing GO terms in column 1 and term names in column 2
}
\description{
This function retrieves the GO term names and IDs of the latest Gene Ontology annotations from Ensembl
}
\details{
This function retrieves the GO term names and IDs of the latest Gene Ontology annotations from Ensembl
}
\examples{
go_names <- get_GO_names('hsapiens')
head(go_names)
}

