% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{get_homology_matrix}
\alias{get_homology_matrix}
\title{get_homology_matrix}
\usage{
get_homology_matrix(species1, species2, seq.identity = FALSE)
}
\arguments{
\item{species1}{Species 1 name, in the form 'hsapiens'}

\item{species2}{Species 2 name, in the form 'hsapiens'}

\item{seq.identity}{Boolean of whether or not  to retrieve sequence identity information, defaults to FALSE}
}
\value{
This function returns a sparse matrix where rows represent the genes in species 1 and columns the genes in species 2.
}
\description{
This function creates a homology matrix between two species by retrieving the homology table and then converting it into a sparse matrix.
}
\details{
This function creates a homology matrix between two species by retrieving the homology table and then converting it into a sparse matrix.
}
\examples{
human_zebrafish_matix <- get_homology_matrix('hsapiens','drerio')
print(data.matrix(human_zebrafish_matix[1:5,1:5]))
}

