% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{get_homology_table}
\alias{get_homology_table}
\title{get_homology_table}
\usage{
get_homology_table(species1, species2, sequence_identity_reference = 0)
}
\arguments{
\item{species1}{Species 1 name, in the form 'hsapiens'}

\item{species2}{Species 2 name, in the form 'hsapiens'}

\item{sequence_identity_reference}{Flag of which sequence identity value to retrieve. 0 (default) returns no value); 1 returns the sequence of identity of genes in species 2 compared to species 1; 2 returns the sequence of identity of genes in species 1 compared to species 2.}
}
\value{
This function return a data frame with 2 or 3 columns, representing the Ensembl gene IDs in species 1 and 2, and the sequence identity score if requested.
}
\description{
This function uses information from Ensembl to generate a homology table between two species.
}
\details{
This function uses information from Ensembl to generate a homology table between two species.
}
\examples{
human_zebrafish_homology_table <- get_homology_table('hsapiens', 'drerio')
head(human_zebrafish_homology_table)

}

