% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{get_overlap_genes}
\alias{get_overlap_genes}
\title{get_overlap_genes}
\usage{
get_overlap_genes(row_genes, col_genes, homology_matrix, min = 5, max = 500)
}
\arguments{
\item{row_genes}{Ensembl gene IDs from species A}

\item{col_genes}{Ensembl gene IDs from species B}

\item{homology_matrix}{The homology matrix for species A and B}

\item{min}{Minimum gene set sizes to be considered, default is 5}

\item{max}{Maximum gene set sizes to be considered, default is 500}
}
\value{
This function returns a list with two vectors of gene IDs, matching the row and column species of the homology table.
}
\description{
This function returns the genes that overlap between the two sets.
}
\details{
This function returns the genes that overlap between the two sets.
}
\examples{
Used within XGSA framework
}

