% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{new_XGSA_dataset}
\alias{new_XGSA_dataset}
\title{new_XGSA_dataset}
\usage{
new_XGSA_dataset(species, data, type = "genesetlist", name,
  universe = "empty")
}
\arguments{
\item{species}{Species name, in the form 'hsapiens'}

\item{data}{A named list of gene sets where each element contains a character vector of Ensembl gene IDs. This can be a named list of only a single element containing gene IDs.}

\item{type}{The type of data provided, currently only 'genesetlist' is supported.}

\item{name}{A name for the data set / collection of gene sets}

\item{universe}{The gene universe for the gene sets in data. They must all have the same gene universe. If no gene universe is provided (default) all Ensembl gene IDs for the indicated species will be used as the gene universe.}
}
\value{
This function returns a list with 5 elements describing the data set, the species, gene sets, data type, name of data set and the applicable gene universe.
}
\description{
This function creates an XGSA data set object that is required for performing tests.
}
\details{
This function creates an XGSA data set object that is required for performing tests.
}
\examples{
test_mouse_genes <- c("ENSMUSG00000033837", "ENSMUSG00000031965", "ENSMUSG00000053110", "ENSMUSG00000030557", "ENSMUSG00000051159", "ENSMUSG00000040289", "ENSMUSG00000035458", "ENSMUSG00000028780", "ENSMUSG00000062327", "ENSMUSG00000037868")
test_mouse_dataset <- new_XGSA_dataset(species = 'mmusculus', data = list(test_genes = test_mouse_genes), name = 'Test Mouse Genes')
print(summary(test_mouse_dataset))
}

