% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\name{paired_fishers_exact_tests}
\alias{paired_fishers_exact_tests}
\title{paired_fishers_exact_tests}
\usage{
paired_fishers_exact_tests(row_genes, col_genes, homology_matrix, min = 5,
  max = 500, rowuniverse, coluniverse)
}
\arguments{
\item{row_genes}{Ensembl gene IDs from species A}

\item{col_genes}{Ensembl gene IDs from species B}

\item{homology_matrix}{The homology matrix for species A and B}

\item{min}{Minimum gene set sizes to be considered, default is 5}

\item{max}{Maximum gene set sizes to be considered, default is 500}

\item{rowuniverse}{Gene universe in species A}

\item{coluniverse}{Gene universe in species B}
}
\value{
This function returns a single numeric p-value
}
\description{
This function executes the paired Fisher's exact tests which form the basis of XGSA.
}
\details{
This function executes the paired Fisher's exact tests which form the basis of XGSA.
}
\examples{
Used within XGSA framework
}

