% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGSA.R
\docType{package}
\name{xgsa-package}
\alias{XGSA}
\alias{xgsa}
\alias{xgsa-package}
\title{XGSA: Cross Species Gene Set Analysis}
\description{
XGSA is a statistical method for cross-species gene set analysis. This package contains the neccessary R code to run XGSA as described in "XGSA: a statistical method for cross-species gene set analysis". XGSA is implemented within a helpful framework to speed up cross-species analyses.
}
\examples{
## Do a sample analyses using cardiac data.


}
\author{
Djordje Djordjevic <D.djordjevic@victorchang.edu.au>
}

